Reverted back the change on BigIncBins in favor of using `prot`
Removed all the BigIncBins related code because all sprites can (and should) use asar to remove their graphics files using `prot {label}` at the *start* of the file, which already effectively does what the BigIncBin macro was doing *and* is simpler *and* doesn't require Pixi doing extra work. Additionally, also fixed a minor error on fread with swapped arguments.

+++ src/file_io.cpp
@@ -12,7 +12,7 @@ FILE *open(const char *name, const char *mode) {
 
 size_t file_size(FILE *file) {
 	fseek(file, 0, SEEK_END);
-	long size = ftell(file);
+	size_t size = ftell(file);
 	fseek(file, 0, SEEK_SET);
 	return size;
 }
